Updated TestLambdaLogger Log method
Updated TestLambdaLogger Log method to perform consistently with the actual Lambda Logger.

Previously, TestLambdaLogger's Log method does the same as the LogLine method (i.e. it appends a newline after each log message). This leads the developer to believe that 
this is how it works in the Lambda Logger. This is not the case though; the Log method in the Lambda Logger does NOT append a newline terminator. If the developer were to assume the behavior was the same and use the Log method instead of LogLine, then things could break.

My personal example, I was validating that my code logs messages in EMF JSON format to post metrics to CloudWatch. I used the Log method (since they appeared the same in TestLambdaLogger) but when I deployed to Lambda, the metrics weren't being sent properly, as all my logs were written on a single line. I had to update all my code to use LogLine.