feat: return 400 on failing to convert target data type for route, query, header or body param
In the current implementation, if the target data type fails to convert, it throws an runtime exception which is not ideal. This runtime failure can handled better by capturing all the conversation failures and return a 400 response with a detailed messages.

This change introduces safe conversion of route, query, header or body param. If any param fails to convert, it is collected and returned in a 400 API Gateway response.