fix: Task being added second time on a 0-event Lambda function which returns a Task type
The issue was due to always wrapping return type to Task when original method is async. This doesn't work for Lambda functions which has 0-events, because their return type can be a Task which is simply delegated to the Lambda runtime without any return type conversion.

Ex. for a API Gateway based Lambda, return type always be API Gateway response type or Task`1
but in case of 0-event Lambda, original method return type can be Task`1 which is again wrapped in Task`1 in T4 template.

This change changes how return type is generated and now ReturnType property of generated method solely controls the code generation. When an API Gateway Lambda return a Task`1, it creates a generic response type, otherwise non-generic response type.

For 0-event based Lambda, there is no action required because their types are already have all the required information.