ci: allow triggering Source Generator CI on workflow_dispatch
Currently there is not a good way to check customer PRs because AWS creds are only available in the main repo and forked repos can't access them (which is intended).

This change will allow developers manually run the check on a PR by pulling into the main repo and trigger it manually.