Fixed issue with incorrect .NET version number being written to runtime-release
Moved the creation of the runtime-release file to the last stage instead of the Lambda Runtime client compilation stage. In the compilation stage the Microsoft provided build base image was used which overrode the ASPNET_VERSION environment variable to the latest ignoring the environment variable we set in the Dockerfile.